Fix built-in scans when buildscript block has apply statements
Built-in scans were breaking with a common pattern where a
script is applied inside the buildscript block, usually in order
to provide a list of commonly used versions for usage inside the block.

This is now fixed by only trying to apply plugin requests when the
script target supports the plugins block.